@@ -1,9 +1,38 @@
 module github.com/zeromicro/goctl-android
 
-go 1.15
+go 1.21
+
+toolchain go1.23.0
 
 require (
-	github.com/stretchr/testify v1.6.1
-	github.com/tal-tech/go-zero v1.1.0
-	github.com/urfave/cli/v2 v2.3.0
+	github.com/samber/lo v1.49.1
+	github.com/stretchr/testify v1.10.0
+	github.com/urfave/cli/v2 v2.11.0
+	github.com/zeromicro/go-zero v1.8.0
+	github.com/zeromicro/go-zero/tools/goctl v1.7.6
+)
+
+require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/fatih/color v1.18.0 // indirect
+	github.com/fatih/structtag v1.2.0 // indirect
+	github.com/gookit/color v1.5.4 // indirect
+	github.com/kr/pretty v0.3.1 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rogpeppe/go-internal v1.10.0 // indirect
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/spaolacci/murmur3 v1.1.0 // indirect
+	github.com/xo/terminfo v0.0.0-20210125001918-ca9a967f8778 // indirect
+	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+	github.com/zeromicro/antlr v0.0.1 // indirect
+	go.opentelemetry.io/otel v1.24.0 // indirect
+	go.opentelemetry.io/otel/trace v1.24.0 // indirect
+	go.uber.org/automaxprocs v1.6.0 // indirect
+	golang.org/x/sys v0.29.0 // indirect
+	golang.org/x/text v0.21.0 // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
