@@ -17,7 +17,8 @@ package generate
 import (
 	"path/filepath"
 
-	"github.com/tal-tech/go-zero/tools/goctl/util"
+	"github.com/zeromicro/go-zero/tools/goctl/util"
+	"github.com/zeromicro/go-zero/tools/goctl/util/pathx"
 	"github.com/zeromicro/goctl-android/template"
 )
 
@@ -49,7 +50,7 @@ type (
 func generateIService(dir string, data IService) error {
 	filename := filepath.Join(dir, "IService.java")
 	base := filepath.Dir(filename)
-	err := util.MkdirIfNotExist(base)
+	err := pathx.MkdirIfNotExist(base)
 	if err != nil {
 		return err
 	}
